move attls related classes to common module
