+++ apiml-tomcat-common/src/main/java/org/zowe/apiml/filter/SecureConnectionFilter.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.apiml.product.filter;
+package org.zowe.apiml.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -17,22 +17,9 @@ import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.StandardCopyOption;
 
 public class SecureConnectionFilter extends OncePerRequestFilter {
-
-    static {
-        try {
-            extractLib(System.getProperty("java.library.path"), "lib" + AttlsContext.ATTLS_LIBRARY_NAME + ".so");
-        } catch (IOException e) {
-            System.out.println(e);
-        }
-    }
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
@@ -45,28 +32,11 @@ public class SecureConnectionFilter extends OncePerRequestFilter {
                 } else {
                     filterChain.doFilter(request, response);
                 }
-            } catch (ContextIsNotInitializedException e) {
-                e.printStackTrace();
-            } catch (UnknownEnumValueException e) {
-                e.printStackTrace();
-            } catch (IoctlCallException e) {
-                e.printStackTrace();
+            } catch (ContextIsNotInitializedException | UnknownEnumValueException | IoctlCallException e) {
+                throw new RuntimeException("Connection security can't be verified", e);
             }
         } else {
             filterChain.doFilter(request, response);
-        }
-    }
-
-    public static void extractLib(String directory, String fileName) throws IOException {
-        File library = new File(directory, fileName);
-        try (InputStream inputStream = AttlsContext.class.getResourceAsStream("/lib/" + fileName)) {
-            Files.copy(inputStream, library.toPath(), StandardCopyOption.REPLACE_EXISTING);
-        } catch (IOException e) {
-            library.delete();
-            throw e;
-        } catch (NullPointerException e) {
-            library.delete();
-            throw new FileNotFoundException(fileName + " does not exist in JAR.");
         }
     }
 }
